chore: move interpreter to separate folder
