+++ src/index.ts
@@ -1,7 +1,7 @@
 export * from "./Parser";
-export * from "./Interpreter";
+export * from "./interpreter/Interpreter";
 export * from "./Tokenizer";
-export * from "./CallableFunction";
+export * from "./interpreter/CallableFunction";
 export * from "./ASTFactories";
 export * from "./typings";
-export * from "./Environment";
+export * from "./interpreter/Environment";
